+++ src/com/iffi/entity/Stock.java
@@ -82,7 +82,7 @@ public class Stock extends Asset {
     }
 
     public double getGainPercentage() {
-        return (double) Math.round(((this.getGain() / this.getPurchaseValue() * 100)) * 1000) / 1000;
+        return Math.round(((this.getGain() / this.getPurchaseValue() * 100)) * 1000.0) / 1000.0;
     }
 
     public String toString() {
